EDIT CERTIFICATE PAGE AND DARK MODE STYLING

+++ src/app/certificate-list/certificate-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Certificate } from '../certificate/certificate';
 import { CertificateService } from '../certificate/certificate.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-certificate-list',
@@ -9,7 +10,7 @@ import { CertificateService } from '../certificate/certificate.service';
 })
 export class CertificateListComponent implements OnInit {
 
-  constructor(private certificateService: CertificateService) { }
+  constructor(private certificateService: CertificateService, private router: Router) { }
   certificates: Certificate[];
 
   ngOnInit(): void {
@@ -28,12 +29,21 @@ export class CertificateListComponent implements OnInit {
   }
 
   deleteCertificate(certificate: Certificate) {
-    this.certificateService.deleteCertificate(certificate.certificateID).subscribe(
-      result => {
-        let index = this.certificates.indexOf(certificate);
-        this.certificates.splice(index, 1);
-      }
-    )
+    if (confirm("Are you sure you want to delete this certificate?")) {
+      this.certificateService.deleteCertificate(certificate.certificateID).subscribe(
+        result => {
+          let index = this.certificates.indexOf(certificate);
+          this.certificates.splice(index, 1);
+        },
+        error => {
+          alert("An error has eccurred");
+        }
+      )
+    }
+  }
+
+  editCertificate(certificateID: number) {
+    this.router.navigate(['/edit-certificate', certificateID])
   }
 
 }
